refactor : Artist 관련 컴포넌트 분리
