+++ feature/home/src/main/java/com/wepli/home/screen/relaylist/ArtistComponent.kt
@@ -1,0 +1,47 @@
+package com.wepli.home.screen.relaylist
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.wepli.feature.home.R
+import com.wepli.shared.feature.mock.artistMockData
+import com.wepli.shared.feature.uimodel.artist.ArtistUiData
+import custom.ArtistProfileListItem
+import custom.TwoLineTitle
+
+
+@Preview
+@Composable
+fun ArtistLayoutPreview() {
+    ArtistLayout(artistList = artistMockData)
+}
+
+@Composable
+fun ArtistLayout(artistList: List<ArtistUiData>) {
+    if (artistList.isEmpty()) return
+
+    Column {
+        TwoLineTitle(
+            title = stringResource(R.string.home_top_artist_title),
+            subscription = stringResource(R.string.home_top_artist_desc),
+        )
+        Spacer(modifier = Modifier.height(4.dp))
+        LazyRow(
+            horizontalArrangement = Arrangement.spacedBy(8.dp),
+            contentPadding = PaddingValues(horizontal = 20.dp)
+        ) {
+            items(artistList) { artist ->
+                ArtistProfileListItem(artist)
+            }
+        }
+    }
+}
